# to_digit simplification (less jumps)