@klensy 2021-05-24
There bultin diff function in godbolt: https://godbolt.org/z/jMsv6dKh8