@gilescope 2021-05-24
Good point. Have flicked the code around so the red / green works in the diff and optimised the to upper case:

https://godbolt.org/z/79KrbxMdY